FilteringData update

+++ Local-B2C/src/main/java/com/b2c/Local/B2C/securities/service/FilteringDataService.java
@@ -1,8 +1,10 @@
 package com.b2c.Local.B2C.securities.service;
 
 import com.b2c.Local.B2C.auths.model.User;
+import com.b2c.Local.B2C.exception.BadRequest400Exception;
 import com.b2c.Local.B2C.exception.NotFound404Exception;
 import com.b2c.Local.B2C.securities.dao.FilterRequestsRepository;
+import com.b2c.Local.B2C.securities.dto.FilterRequestInfo;
 import com.b2c.Local.B2C.securities.model.FilterRequest;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.core.context.SecurityContextHolder;
@@ -10,6 +12,7 @@ import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.session.FindByIndexNameSessionRepository;
 import org.springframework.session.Session;
 import org.springframework.stereotype.Service;
+
 import java.util.List;
 import java.util.Set;
 
@@ -34,7 +37,7 @@ public class FilteringDataService {
         return null;
     }
 
-    public List<FilterRequest> getAllSessionId() {
+    public List<FilterRequest> getAllSessionRequests() {
         if (getLoggedInUserId() != null)
             return filterRequestsRepository.findAllByUserId(getLoggedInUserId().getId().toString());
         else {
@@ -59,9 +62,29 @@ public class FilteringDataService {
     }
 
     public void destroyAllSession() {
-        getActiveSessionId().forEach(s -> {
-            sessions.deleteById(s);
-        });
+        if (getActiveSessionId().size() != 0) {
+            getActiveSessionId().forEach(s -> {
+                sessions.deleteById(s);
+            });
+        } else {
+            throw new BadRequest400Exception("Not Any Active Session");
+        }
+    }
+
+    public List<FilterRequest> getRequestsBySessionId(String sessionId, String url) {
+        if (filterRequestsRepository.existsAllByUrlAndSessionId(url, sessionId)) {
+            return filterRequestsRepository.findAllByUrlAndSessionId(url, sessionId);
+        } else {
+            throw new NotFound404Exception("Not Found");
+        }
+    }
+
+    public List<FilterRequestInfo> getAllSessionsByUser() {
+        if (getLoggedInUserId() != null)
+            return filterRequestsRepository.getAllByUserId(getLoggedInUserId().getId().toString());
+        else {
+            throw new NotFound404Exception("Not Found");
+        }
     }
 }
